Improve status_reason when cluster creation is stalled
If during cluster creation or update, one or more nodes cannot be
created by Nova (because, for example, the project runs out of
quota, or Nova runs out of suitable hypervisors), the cluster stalls
with CREATE_IN_PROGRESS status. The status_reason in this case is
not helpful as it only reports the previous step which succeeded.

This patch adds a check of each individual machine status looking
for machines which are not ready and reports any reason found.
This can result in useful status_reason messages such as
 'error creating Openstack instance ... Quota exceeded for instances'